requirements.txt
python-hcl2==7.3.1
requests==2.32.5